# src/main/java/com/hafricker/gen/api/IbansApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.hafricker.gen.api;

import com.hafricker.gen.dto.Error;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;
import springfox.documentation.annotations.ApiIgnore;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-08-31T22:46:55.341766800+02:00[Europe/Paris]")
@Validated
@Api(value = "ibans", description = "the ibans API")
public interface IbansApi {

    /**
     * GET /ibans/{iban}
     * check if an Iban is valid
     *
     * @param iban The iban to validate (required)
     * @return the Iban is valid (status code 200)
     *         or unexpected error (status code 200)
     */
    @ApiOperation(value = "", nickname = "checkIban", notes = "check if an Iban is valid", tags={ "IbanApi", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "the Iban is valid"),
        @ApiResponse(code = 200, message = "unexpected error", response = Error.class) })
    @GetMapping(
        value = "/ibans/{iban}",
        produces = { "application/json" }
    )
    ResponseEntity<Void> checkIban(@ApiParam(value = "The iban to validate",required=true) @PathVariable("iban") String iban);

}