# src/main/resources/swagger-bank.yaml
openapi: "3.0.0"
info:
  title: Bank Validator swagger
  description: This is the api documentation of bank validator.
  termsOfService: http://bankvalidator.com/terms/
  contact:
    name: API Support
    url: http://www.bankvalidator.com/support
    email: support@hafricker.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://bankvalidator.com/{basePath}
    description: The production API server
    variables:
      basePath:
        # open meaning there is the opportunity to use special base paths as assigned by the provider, default is `v1`
        default: v1
paths:
  /countries:
    get:
      tags:
        - CountryApi
      description: "Get countries within iban validator"
      operationId: "getCountries"
      responses:
        '200':
          description: A list of Country.
          headers:
            x-count:
              description: number of items return
              schema:
                type: integer
                format: int64
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
  /ibans/{iban}:
    get:
      tags:
        - IbanApi
      description: "check if an Iban is valid"
      operationId: "checkIban"
      parameters:
        - name: iban
          in: path
          required: true
          description: The iban to validate
          schema:
            type: string
      responses:
        '200':
          description: the Iban is valid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    Country:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        continent:
          type: string
        code:
          type: string